Updated deps; API changes for model downloader API; renamed ModelDownloader -> ModelLoader to disambiguate with underlying API method
