+++ final/ContentRecommendations/AllMoviesCollectionViewController.swift
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import UIKit
-import Firebase
+import FirebaseAnalytics
 
 class AllMoviesCollectionViewController: UICollectionViewController {
